refactor: move credit card summary collection to credit-cards
