+++ apps/web/src/integrations/tanstack-db/credit-cards.ts
@@ -338,3 +338,43 @@ export function bulkCreateCreditCardsAction(collection: CreditCardsCollection) {
       },
    });
 }
+
+export type CreditCardSummaryCollectionRow =
+   Outputs["transactions"]["getSummary"] & {
+      id: string;
+      creditCardId: string;
+      dateFrom: string;
+      dateTo: string;
+   };
+
+type CreditCardSummaryCollectionOptionsParams = {
+   queryClient: QueryClient;
+   creditCardId: string;
+   dateFrom: string;
+   dateTo: string;
+};
+
+export function creditCardSummaryCollectionOptions({
+   queryClient,
+   creditCardId,
+   dateFrom,
+   dateTo,
+}: CreditCardSummaryCollectionOptionsParams) {
+   const id = `${creditCardId}:${dateFrom}:${dateTo}`;
+   return queryCollectionOptions({
+      id: `credit-card-summary:${id}`,
+      queryKey: () => ["credit-card-summary", id],
+      queryFn: async () => {
+         const summary = await orpc.transactions.getSummary.call({
+            creditCardId,
+            dateFrom,
+            dateTo,
+         });
+         return [{ ...summary, id, creditCardId, dateFrom, dateTo }];
+      },
+      queryClient,
+      getKey: (summary: CreditCardSummaryCollectionRow) => summary.id,
+      refetchInterval: 5_000,
+      syncMode: "on-demand",
+   });
+}
